# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API REST de videojuegos
  description: API REST para consultar y gestionar videojuegos, desarrolladores, reviews y países relacionados con sus desarrolladores.
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Servidor local de desarrollo
...
paths:
...
  /developers:
    get:
      tags:
        - Developers
      summary: Obtener desarrolladores
      description: Lista los desarrolladores almacenados en MongoDB. Permite búsqueda, filtro por videojuego, ordenación y paginación.
      parameters:
        - name: search
          in: query
          description: Texto a buscar en nombre o slug.
          schema:
            type: string
          example: Valve
        - name: gameId
          in: query
          description: ID de un videojuego asociado al desarrollador.
          schema:
            type: integer
            minimum: 1
          example: 3328
        - name: page
          in: query
          description: Número de página.
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 2
        - name: limit
          in: query
          description: Número máximo de resultados.
          schema:
            type: integer
            minimum: 1
            default: 1000
          example: 20
        - name: sort
          in: query
          description: "Campo de ordenación. Con guion inicial ordena de forma descendente. Campos permitidos: id, name, slug, games_count."
          schema:
            type: string
          example: -games_count
      responses:
        "200":
          description: Lista de desarrolladores.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DevelopersListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - Developers
      summary: Crear desarrollador
      description: Inserta un desarrollador en la colección de desarrolladores.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeveloperInput"
      responses:
        "201":
          description: Desarrollador creado correctamente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedDeveloperResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Ya existe un desarrollador con ese ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /developers/{id}:
    get:
      tags:
        - Developers
      summary: Obtener desarrollador por ID
      description: Busca un desarrollador por el ID numérico importado desde RAWG.
      parameters:
        - name: id
          in: path
          required: true
          description: ID numérico del desarrollador.
          schema:
            type: integer
            minimum: 1
          example: 1612
      responses:
        "200":
          description: Desarrollador encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Developer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      tags:
        - Developers
      summary: Actualizar desarrollador
      description: Actualiza los campos admitidos de un desarrollador existente.
      parameters:
        - name: id
          in: path
          required: true
          description: ID numérico del desarrollador.
          schema:
            type: integer
            minimum: 1
          example: 1612
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeveloperUpdate"
      responses:
        "200":
          description: Desarrollador actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedDeveloperResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags:
        - Developers
      summary: Eliminar desarrollador
      description: Elimina un desarrollador de la colección.
      parameters:
        - name: id
          in: path
          required: true
          description: ID numérico del desarrollador.
          schema:
            type: integer
            minimum: 1
          example: 1612
      responses:
        "200":
          description: Desarrollador eliminado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  responses:
    BadRequest:
      description: Petición incorrecta.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Error interno del servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DeveloperGame:
      type: object
      properties:
        id:
          type: integer
          example: 4200
        slug:
          type: string
          example: portal-2
        name:
          type: string
          example: Portal 2
        added:
          type: integer
          example: 20858
    Developer:
      type: object
      properties:
        id:
          type: integer
          example: 1612
        name:
          type: string
          example: Valve Software
        slug:
          type: string
          example: valve-software
        games_count:
          type: integer
          example: 44
        image_background:
          type: string
          nullable: true
        games:
          type: array
          items:
            $ref: "#/components/schemas/DeveloperGame"
    DeveloperInput:
      type: object
      required:
        - id
        - name
        - slug
        - games
      properties:
        id:
          type: integer
          example: 999999
        name:
          type: string
          example: Mi Estudio
        slug:
          type: string
          example: mi-estudio
        games_count:
          type: integer
          example: 1
        image_background:
          type: string
          nullable: true
          example: https://example.com/developer.jpg
        games:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/DeveloperGame"
    DeveloperUpdate:
      type: object
      description: Campos del desarrollador que se quieran modificar. El ID no se puede actualizar.
      properties:
        name:
          type: string
        slug:
          type: string
        games_count:
          type: integer
        image_background:
          type: string
          nullable: true
        games:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/DeveloperGame"
...
    DevelopersListResponse:
      type: object
      properties:
        total:
          type: integer
          example: 600
        developers_length:
          type: integer
          example: 2
        developers:
          type: array
          items:
            $ref: "#/components/schemas/Developer"
...
    CreatedDeveloperResponse:
      type: object
      properties:
        message:
          type: string
          example: Desarrollador creado correctamente
        id:
          type: integer
          example: 7003
        developer:
          $ref: "#/components/schemas/Developer"
    UpdatedDeveloperResponse:
      type: object
      properties:
        message:
          type: string
          example: Desarrollador actualizado correctamente
        developer:
          $ref: "#/components/schemas/Developer"
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: El id debe ser un número positivo
        error:
          type: object
          description: Detalle técnico cuando el error procede de una excepción.
          additionalProperties: true
          nullable: true
        missingFields:
          type: array
          description: Campos obligatorios que faltan en la petición.
          items:
            type: string
        invalidFields:
          type: array
          description: Campos enviados que la ruta no admite.
          items:
            type: string